requirements.txt
pyTelegramBotAPI==4.14.0
pynotiq-client>=0.3.0
python-dotenv==1.0.0